+++ app/swap/page.tsx
@@ -1,11 +1,21 @@
 import Header from "@/components/ui/header";
 import { SwapPanel } from "./SwapPanel";
 
-export default function SwapPage() {
+type SwapPageProps = {
+  searchParams?: {
+    tokenId?: string;
+    tokenName?: string;
+  };
+};
+
+export default function SwapPage({ searchParams }: SwapPageProps) {
   return (
     <div className="min-h-screen flex flex-col">
       <Header />
-      <SwapPanel />
+      <SwapPanel
+        initialTokenId={searchParams?.tokenId}
+        initialTokenName={searchParams?.tokenName}
+      />
     </div>
   );
 }